Sept 28 to Oct 2 work
Fortran implementation is not working yet. I can have the C++ CalcSpecConc function work or the Fortran, but not both. Working on trying to tie in the Fortran code with Rcpp package instead of using a call to .Fortran in an R script. FKellyFunctions.R has just been moved to /scrap for now.