first cut of expandWHAM function
The wDLF, wKZED, wKsel, wP, wRadius, and wMolWt parameters are all passed along for later calculations, and Comp and Spec parameters are modified for inclusion of WHAM components and reactions, but the DOC component is not removed. The dLK1A, dLK1B, and dLK2 parameters are read in but not currently do not do anything, since they were constant/not used in Model V, but were added in Model VI and VII. We need to figure out where to put these - either passed along or used within the expandWHAM function. Linked Lists and Phases also need to be added here, as they will need the number of components expanded in their arrays.