pH needs a defined component now
I was initially thinking of making the "From" column say something like "10^-pH", but didn't feel right about making it seem like it's an evaluated equation. Next I was thinking "antilog(pH)", but that is unclear for the reasons that (1) it's not explicit about it being log10 vs natural log, (2) it's excluding the detail of needing the negative sign, and (3) it makes it seem like there should also be a "log(Var)" option. The whole thing was getting complicated, so I simplified to to just take the approach we did with the input variable - pH is just pH and you use it consistent with how you should always treat pH. There's no guarantee we would ever need to do a log/antilog conversion for any other reason, and the point is to be able to specify information about activity and FixedConc vs FixedAct, so let's just leave it at that. K.I.S.S.