WHAM revamp and file convert functions
WHAM now acts more like its own problem object. The "WHAM" list item in a BlankProblem() type list is basically a plug-and-play item that has all of the information defined in a wdat file and everything needed by ExpandWHAM().  This change is partially in response to having ConvertWHAMThermoFile() and ConvertWindowsParamFile() functions. The Windows parameter files and thermo files have subtle differences beyond just being for "WHAM V". These differences need to be captured in the problem objects.